XNAT-2241: Merged counts, made per-format display a secondary pop-up text. Added regen function to Manage Files, forced page reload for resource updates.
Former-commit-id: 36125caaf383c26ede9ef910dc7bdf7743d39764

+++ plugin-resources/webapp/xnat/java/org/nrg/xdat/om/base/BaseXnatAbstractresource.java
@@ -72,7 +72,7 @@ public abstract class BaseXnatAbstractresource extends AutoXnatAbstractresource 
         if (size ==null){
             calculate(rootPath);
         }
-        return size.longValue();
+        return size;
     }
     
     public String getReadableFileStats() {
